refactor: adjust android sample
