+++ sample/android/src/main/java/co/yml/ychat/android/presentation/completions/CompletionsScreen.kt
@@ -9,6 +9,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
+import co.yml.openai.provider.OpenAi
 import co.yml.ychat.YChat
 import co.yml.ychat.android.BuildConfig
 import co.yml.ychat.android.R
@@ -50,8 +51,8 @@ internal fun CompletionsScreen(viewModel: CompletionsViewModel = getViewModel())
 @Composable
 private fun CompletionsScreenPreview() {
     YChatTheme {
-        val yChat = YChat.create(BuildConfig.API_KEY)
-        val viewModel = CompletionsViewModel(yChat)
+        val openAi = OpenAi.create(BuildConfig.API_KEY)
+        val viewModel = CompletionsViewModel(openAi)
         CompletionsScreen(viewModel)
     }
 }
