+++ sample/android/src/main/java/co/yml/ychat/android/presentation/edits/EditsScreen.kt
@@ -11,6 +11,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import co.yml.openai.provider.OpenAi
 import co.yml.ychat.YChat
 import co.yml.ychat.android.BuildConfig
 import co.yml.ychat.android.R
@@ -68,8 +69,8 @@ internal fun EditsScreen(viewModel: EditsViewModel = getViewModel()) {
 @Composable
 private fun EditsScreenPreview() {
     YChatTheme {
-        val yChat = YChat.create(BuildConfig.API_KEY)
-        val viewModel = EditsViewModel(yChat)
+        val openAi = OpenAi.create(BuildConfig.API_KEY)
+        val viewModel = EditsViewModel(openAi)
         EditsScreen(viewModel)
     }
 }
