+++ sample/android/src/main/java/co/yml/ychat/android/presentation/models/viewmodel/ModelsViewModel.kt
@@ -2,16 +2,16 @@ package co.yml.ychat.android.presentation.models.viewmodel
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import co.yml.ychat.YChat
-import co.yml.ychat.domain.model.AIModel
+import co.yml.openai.provider.OpenAi
+import co.yml.openai.provider.domain.model.AIModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
 
-internal class ModelsViewModel(private val yChat: YChat) : ViewModel() {
+internal class ModelsViewModel(private val openAi: OpenAi) : ViewModel() {
 
-    private val yChatListModels by lazy { yChat.listModels() }
+    private val listModels by lazy { openAi.listModels() }
 
     private val _state = MutableStateFlow<State>(State.Loading)
     val state: StateFlow<State> = _state.asStateFlow()
@@ -22,7 +22,7 @@ internal class ModelsViewModel(private val yChat: YChat) : ViewModel() {
 
     fun fetchModels() = viewModelScope.launch {
         _state.value = State.Loading
-        runCatching { yChatListModels.execute() }
+        runCatching { listModels.execute() }
             .onSuccess { _state.value = State.Success(it) }
             .onFailure { _state.value = State.Error }
     }
